# doc/api.yaml
openapi: 3.0.0
paths:
...
  /album:
    get:
      operationId: AlbumsController_find
      summary: Get albums list
      description: Gets all library albums list
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        '401':
          description: Access token is missing or invalid
      tags:
        - Album
    post:
      operationId: AlbumsController_create
      summary: Add new album
      description: Add new album information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAlbumDto'
      responses:
        '201':
          description: Album is created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. body does not contain required fields
        '401':
          description: Access token is missing or invalid
      tags:
        - Album
  /album/{id}:
    get:
      operationId: AlbumsController_findOne
      summary: Get single album by id
      description: Get single album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found.
      tags:
        - Album
    put:
      operationId: AlbumsController_update
      summary: Update album information
      description: Update library album information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found
      tags:
        - Album
    delete:
      operationId: AlbumsController_delete
      summary: Delete album
      description: Delete album from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been deleted
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Album not found
      tags:
        - Album
...
  /artist:
    get:
      operationId: ArtistsController_find
      summary: Gets all artists
      description: Gets all artists
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '401':
          description: Access token is missing or invalid
      tags:
        - Arrtist
    post:
      operationId: ArtistsController_create
      summary: Add new artist
      description: Add new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddArtistDto'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. body does not contain required fields
        '401':
          description: Access token is missing or invalid
      tags:
        - Arrtist
  /artist/{id}:
    get:
      operationId: ArtistsController_findOne
      summary: Get single artist
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found.
      tags:
        - Arrtist
    put:
      operationId: ArtistsController_update
      summary: Update artist information
      description: Update library artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The artist has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found
      tags:
        - Arrtist
    delete:
      operationId: ArtistsController_delete
      summary: Delete artist
      description: Delete artist from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '401':
          description: Access token is missing or invalid
        '404':
          description: Artist not found
      tags:
        - Arrtist
...
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      summary: Add album to favorites
      description: Add album to favorite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFav'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '422':
          description: Album not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteAlbum
      summary: Delete album from favorites
      description: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been deleted
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to favorites
      description: Add artist to favorite
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFav'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '422':
          description: Artist not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteArtist
      summary: Delete artist from favorites
      description: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Favorites
info:
  title: REST Service
  description: >-
    Let's try to create a Home Library Service! Users can create, read, update,
    delete data about Artists, Tracks and Albums, add them to Favorites in their
    own Home Library!
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
...
    Album:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: Music for the Jilted Generation
        year:
          type: number
          example: 1994
        artistId:
          type: string
          example: d2992f19-7cac-4a86-9b3d-9d143a42b66d
      required:
        - id
        - name
        - year
        - artistId
    AddAlbumDto:
      type: object
      properties:
        name:
          type: string
          default: Dynamite
        year:
          type: number
          default: 2005
        artistId:
          type: string
          default: null
      required:
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          default: null
      required:
        - name
        - year
...
    Artist:
      type: object
      properties:
        id:
          type: string
          example: 97c8d54a-b5e4-4a71-90b6-3786a5bbc094
        name:
          type: string
          example: Prodigy
        grammy:
          type: boolean
          example: true
      required:
        - id
        - name
        - grammy
    AddArtistDto:
      type: object
      properties:
        name:
          type: string
          default: Jamiroquai
        grammy:
          type: boolean
          default: true
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
...
    ResponseFav:
      type: object
      properties:
        message:
          type: string
          example:
            message: Track successfully added
      required:
        - message